Switching to indirect cloning/mutation with helpers/wrappers.
Switched traverse to using helpers. Switched $slice to using
wrapping objects. This proves out the concept (and performance
which is tricky) of abstracting away the direct cloning and
mutation of objects and arrays. This cleans up the code,
getting rid of some weird "lookahead" bits and "mutation
markers" and paves the way for working with other objects
like Maps and Sets.